pulling out the issuance authorization key material into its own module to separate out the issuance authorization material
