+++ tests/issuance_global_state.rs
@@ -14,7 +14,8 @@ use orchard::{
         },
         IssueBundle, IssueInfo, Signed,
     },
-    keys::{FullViewingKey, IssuanceAuthorizingKey, IssuanceValidatingKey, Scope, SpendingKey},
+    issuance_auth::{IssuanceAuthorizingKey, IssuanceValidatingKey},
+    keys::{FullViewingKey, Scope, SpendingKey},
     note::{AssetBase, Nullifier},
     value::NoteValue,
     Address, Note,
